replace silent ImportError with semantic error message when binding does not exist
