+++ python/perspective/perspective/__init__.py
@@ -6,17 +6,10 @@
 # the Apache License 2.0.  The full license can be found in the LICENSE file.
 #
 
+from .libpsp import *  # noqa: F401, F403
 from .core import *  # noqa: F401, F403
 from .core._version import __version__  # noqa: F401
-from .manager import *  # noqa: F401, F403
-from .tornado_handler import *  # noqa: F401, F403
-from .viewer import *  # noqa: F401, F403
 from .widget import *  # noqa: F401, F403
-
-try:
-    from .table import *  # noqa: F401, F403
-except ImportError:
-    pass
 
 try:
     from .node import *  # noqa: F401, F403
