+++ python/perspective/perspective/table/__init__.py
@@ -7,20 +7,6 @@
 #
 
 from .table import Table
+from .libbinding import PerspectiveCppError
 
-
-def is_libpsp():
-    """Was libbinding successfully loaded in this module?"""
-    return __is_libpsp__
-
-
-__all__ = ["Table", "is_libpsp"]
-
-__is_libpsp__ = True
-
-try:
-    from .libbinding import PerspectiveCppError  # noqa: F401
-    __all__.append("PerspectiveCppError")
-except ImportError:
-    __is_libpsp__ = False
-    pass
+__all__ = ["Table", "PerspectiveCppError"]
