+++ python/perspective/perspective/table/_data_formatter.py
@@ -9,13 +9,9 @@
 import numpy as np
 from math import trunc
 from ._constants import COLUMN_SEPARATOR_STRING
-
-try:
-    from .libbinding import get_data_slice_zero, get_data_slice_one, get_data_slice_two, \
-        get_from_data_slice_zero, get_from_data_slice_one, get_from_data_slice_two, \
-        get_pkeys_from_data_slice_zero, get_pkeys_from_data_slice_one, get_pkeys_from_data_slice_two
-except ImportError:
-    pass
+from .libbinding import get_data_slice_zero, get_data_slice_one, get_data_slice_two, \
+    get_from_data_slice_zero, get_from_data_slice_one, get_from_data_slice_two, \
+    get_pkeys_from_data_slice_zero, get_pkeys_from_data_slice_one, get_pkeys_from_data_slice_two
 
 
 def _mod(a, b):
